pyproject.toml
[project]
name = "accounting-assistant-agent"
version = "0.1.0"
description = "Categorizes transactions"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.0",
    "google-adk>=1.7.0", 
    "python-dotenv>=1.0.0"
]

[tool.setuptools]
packages = ["acc_agent"]
